feat(git): Add git hook to validate commit message
