+++ buildSrc/src/main/resources/tequila/hooks/commit-msg
@@ -1,0 +1,11 @@
+#!/usr/bin/env bash
+
+msg=$(cat "$1")
+result=$(./gradlew validateCommitMessageGit -Pmessage="$msg" --console=plain)
+regex='Error: (.+)BUILD SUCCESSFUL.*'
+
+if [[ $result =~ $regex ]]
+then
+  echo "${BASH_REMATCH[1]}";
+  exit 1;
+fi
