+++ settings.gradle.kts
@@ -1,1 +1,2 @@
 rootProject.name = "CHANGE_THAT"
+include("main")
